requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.1.0
httpie==2.3.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.2
Pillow==7.2.0
psycopg2==2.8.5
Pygments==2.7.4
pylint==2.5.3
PySocks==1.7.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
uuid==1.30
whitenoise==5.2.0
wrapt==1.12.1
